moved base pipeline to separate file
