+++ luxonis_ml/data/augmentations/__init__.py
@@ -1,9 +1,11 @@
 from .augmentations import Augmentations
+from .base_pipeline import BaseAugmentationPipeline
 from .batch_compose import BatchCompose
 from .custom import LetterboxResize, MixUp, Mosaic4
 
 __all__ = [
     "Augmentations",
+    "BaseAugmentationPipeline",
     "LetterboxResize",
     "MixUp",
     "Mosaic4",
